server/services/bias_analysis/requirements.txt
fastapi==0.95.2
uvicorn==0.24.0
python-multipart==0.0.6
pandas==2.1.3
numpy==1.26.2
aif360==0.5.0
scikit-learn==1.3.2
python-dotenv==1.0.0
pydantic==1.10.13
python-magic==0.4.27
aiofiles==23.2.1
openpyxl 